Changed Flag events, FlagCaptured logic.
Added delegate events for player flags changed and opponent flags changed so the UI can update correctly.
Changed reference types to player and opponent.
Added flag capture logic for the beginning of the game when nobody has a flag yet. The first character to capture a flag makes the opponent "IT".